+++ changelog.md
@@ -4,6 +4,13 @@
 
 ---
 
+### v0.9.3
+
+2025-02-09
+
+* Upgrade Menu - Dynamic Mobile Breakpoint
+* FIX - Include No Branding Post Template
+
 ### v0.9.2
 
 2025-02-08
@@ -19,7 +26,7 @@
 
 ### v0.9.0
 
-*No Public Release*
+No Public Release
 
 * Initial reconfiguration of styles and scripts
 * Beta Version development begins